+++ reducer_extension/test/reducer_extension_test.dart
@@ -1,9 +1,9 @@
-import 'package:test/test.dart';
 import 'package:reducer_extension/reducer_extension.dart';
+import 'package:test/test.dart';
 
 void main() {
   test('annotation exists', () {
     final annotation = const SimpleObject();
-    expect(annotation is Reducer, true);
+    expect(annotation is ActionReducer, true);
   });
 }