+++ reducer_extension/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2 Fix name clash with Reducer Class from redux
+
+* change annotation from `@reducer` to `@actionReducer` since it generates both action and reducer.
+
 ## 0.0.1 Initial release
 
 * Lets you annotate an object using `@reducer` to generate a `reducer` extension.